+++ jest.config.js
@@ -15,11 +15,6 @@ module.exports = {
 	setupFilesAfterEnv: [
 		"@testing-library/jest-dom/extend-expect"
 	],
-	preset: "ts-jest/presets/js-with-ts",
-	testEnvironment: "jsdom",
-	globals: {
-		'ts-jest': {
-			tsconfig: 'tsconfig.test.json'
-		}
-	}
+	preset: "ts-jest",
+	testEnvironment: "jsdom"
 };