+++ src/index.ts
@@ -1,2 +1,3 @@
 export { default as usePastDelay } from './usePastDelay';
 export { default as lazy } from './lazy';
+export { default as Suspense } from './Suspense';