+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
+### Added
+- `Suspense` wrapper.
+
 ## [2.0.0][] - 2022-07-31
 
 ### Added